requirements.txt
RPi.GPIO>=0.7.0
spidev>=3.5
Pillow>=9.0.0
psutil>=5.8.0